[NORM] #67 - norminette
